the third major interface revision
the pane now only displays on the foaf.
tweets have a stream view for people you follow,
a personal view that allows you to see your updates
a view of those updates people have mentioned you in
and a follow view.

there are still bugs with the view on other people's foafs
fixes for this will come later.

also there is a bug with the tabs if they are too long for
the space horizontally. this will be fixed soon.